+++ src/DistinctOn/DistinctOnWalkerException.php
@@ -1,0 +1,18 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Copyright (c) 2025-2026 Maksim Tyugaev
+ *
+ * For the full copyright and license information, please view
+ * the LICENSE file that was distributed with this source code.
+ *
+ * @see https://github.com/tugmaks/doctrine-walkers
+ */
+
+namespace Tugmaks\DoctrineWalkers\DistinctOn;
+
+final class DistinctOnWalkerException extends \LogicException
+{
+}